Fix dead code warnings for async/blocking versions
